move gethostname to System
