+++ src/cpp/core/system/PosixSystem.cpp
@@ -3163,39 +3163,6 @@ Error restoreRoot()
    return posix::restoreRoot();
 }
 
-/**
- * Returns the hostname from the operating system
- */
-std::string getHostname()
-{
-   // Use a static string to store the hostname so we don't have to look it up
-   // multiple times
-   static std::string hostname;
-   static boost::mutex mutex;
-   std::string result;
-
-   // Lock to ensure that we don't try to read/write the hostname from two
-   // threads
-   LOCK_MUTEX(mutex)
-   {
-      if (hostname.empty())
-      {
-         char buffer[256];
-         int status = ::gethostname(buffer, 255);
-         if (status == 0)
-         {
-            // If successful, store the hostname for later; swallow errors here
-            // since they are not actionable
-            hostname = std::string(buffer);
-         }
-      }
-      result = hostname;
-   }
-   END_LOCK_MUTEX
-
-   return result;
-}
-
 } // namespace system
 } // namespace core
 } // namespace rstudio
